examples/python/facilitator/requirements.txt
x402>=0.1.0
python-dotenv>=1.0.0
tronpy>=0.4.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
